Upstream's September work: two directory moves (src/melee/ft/chara to
ft/kinds, extern/dolphin to libs/dolphin), the tree-wide change from
<platform.h> to <Runtime/platform.h>, more file splits, and a long list of
field and function renames as more of the game is named.

195 files conflicted, 632 hunks, resolved by taking both sides: upstream's
renames and refactors adopted, the port's host fixes carried onto them, then
checked against the port markers each file carried before.

What went wrong, and what it taught:

- <platform.h> resolved to the port's own header on the host, where SDL, GL
  and __OSBusClock live. The new spelling reaches upstream's header instead,
  so that header now includes the port's; otherwise seven hundred
  translation units quietly lose the port layer.

- Three renamed functions still had callers on the old names, and each old
  name had a weak stub to answer them. The callers follow the renames and
  the stubs are deleted, with six more that upstream's real definitions have
  made dead.

- Upstream expressed the stock HUD's per-player data as arithmetic from the
  struct base: count the player's 0x54-byte element, then add one 0x204-byte
  header. That equals the array only because it sits at 0x204 on the disc.
  Here the array starts at 0x3D8 and the arithmetic landed 468 bytes early,
  inside players 3 to 5's stock-icon joints, so the animation counters were
  read from and written over three players' HUD. A versus match hid it;
  Classic crashed walking joints made of counter bytes.

Two more of that same class were found by auditing for it rather than by
crashing: a 120-byte overrun of the records screen's sorting data, and the
Classic round table, which was indexed and walked with no bound.

Refraction stays off on the host, as before. Golden suite within baseline;
Classic runs to completion on the desktop and six matches on the tablet at
60 fps.

+++ src/MSL/math.h
@@ -6,10 +6,11 @@
 #else
 #define MSL_MATH_H
 
-#include <platform.h>
+#include <Runtime/platform.h>
 
 #ifdef MWERKS_GEKKO
 #include <math_ppc.h>
+
 #include <MetroTRK/intrinsics.h>
 #endif
 
@@ -92,7 +93,7 @@ void __sinit_trigf_c(void);
 
 static inline float fmodf(float a, float b)
 {
-    long long quotient;
+    s64 quotient;
 
     if (fabsf(b) > fabsf(a)) {
         return a;
